Bridge fuckery
Messed around with the custom bridge collisions, finally figured it out. implemented it on the first bridge, second bridge will probably need a second passover. also messed with the disappear bridge 1 bp. got the materials changing. looks better in my opinion. also got the blue working on it, because that was a thing apparently